:recycle: (import): move the upload-batch constants and types to their own file
